RD-4466 deployment update: factor out a module
I'm going to be changing a lot around this, and workflow.py is
already 1k lines long, so let's factor this out.
