+++ workflows/cloudify_system_workflows/deployment_update/utils.py
@@ -1,0 +1,4 @@
+def clear_graph(graph):
+    """Remove all tasks from the graph"""
+    for task in graph.tasks:
+        graph.remove_task(task)
